# gh-95865: Speed up urllib.parse.quote_from_bytes()