--- a/Misc/NEWS.d/next/Library/2022-08-11-03-16-48.gh-issue-95865.0IOkFP.rst
+++ b/Misc/NEWS.d/next/Library/2022-08-11-03-16-48.gh-issue-95865.0IOkFP.rst
@@ -1,0 +1,1 @@
+Speed up :func:`urllib.parse.quote_from_bytes` by replacing a list comprehension with ``map()``.